queda por solucionar fallo de que al guardar juegos eliminarlo, se elimina de la "raiz" y no puedo volver agrgarlo, y ahora los mensajes se eliminan pasado el tiempo que yo quiera

+++ app/src/main/java/com/proyectofinal/redgame/ui/foro/workers/ClearChatWorker.kt
@@ -1,28 +1,38 @@
 package com.proyectofinal.redgame.ui.foro.workers
 
 import android.content.Context
+import android.util.Log
 import androidx.work.CoroutineWorker
 import androidx.work.Worker
 import androidx.work.WorkerParameters
 import com.google.firebase.firestore.FirebaseFirestore
 import kotlinx.coroutines.tasks.await
 
-class ClearChatWorker( context: Context,workerParams: WorkerParameters): CoroutineWorker(context,workerParams){
+class ClearChatWorker( context: Context,workerParams: WorkerParameters): CoroutineWorker(context,workerParams) {
 
     override suspend fun doWork(): Result {
+        Log.d("ClearChatWorker", "Ejecutando ClearChatWorker a las: ${System.currentTimeMillis()}")
+
         val db = FirebaseFirestore.getInstance()
         val chatCollection = db.collection("Posts")
 
         return try {
-            val snapshot = chatCollection.get().await() // Espera a obtener el snapshot
-            for (document in snapshot) {
-                document.reference.delete().await() // Espera a eliminar cada documento
+            // Obtener y borrar los documentos en la colección "Posts"
+            val snapshot = chatCollection.get().await() // Espera la finalización de la tarea
+
+            // Verifica si hay documentos para borrar
+            if (snapshot.isEmpty) {
+                Log.d("ClearChatWorker", "No hay documentos para borrar.")
+            } else {
+                for (document in snapshot.documents) {
+                    document.reference.delete().await() // Espera que se elimine cada documento
+                }
+                Log.d("ClearChatWorker", "Chat borrado exitosamente.")
             }
             Result.success()
         } catch (e: Exception) {
-            e.printStackTrace() // Imprime el error para depuración
-            Result.failure() // Retorna failure si algo sale mal
+            Log.e("ClearChatWorker", "Error en ClearChatWorker: ${e.message}")
+            Result.failure()
         }
     }
-
 }
